requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pyspark==3.1.2
cassandra-driver==3.25.0